markdown: Rename preprocessor_priorities module to priorities.
