+++ zerver/lib/markdown/fenced_code.py
@@ -88,7 +88,7 @@ from pygments.lexers import find_lexer_class_by_name
 from pygments.util import ClassNotFound
 
 from zerver.lib.exceptions import MarkdownRenderingException
-from zerver.lib.markdown.preprocessor_priorities import PREPROCESSOR_PRIORITES
+from zerver.lib.markdown.priorities import PREPROCESSOR_PRIORITES
 from zerver.lib.tex import render_tex
 
 # Global vars
